improve

+++ plannap-android/app/src/main/java/studio/aquatan/plannap/data/model/PostSpot.kt
@@ -8,6 +8,6 @@ data class PostSpot(
     @Json(name = "name") val name: String,
     @Json(name = "note") val note: String,
     @Json(name = "image") val image: String,
-    @Json(name = "lat") val lat: Float,
-    @Json(name = "lon") val lon: Float
+    @Json(name = "lat") val lat: Double,
+    @Json(name = "lon") val lon: Double
 )
